Added file extension actions.

+++ src/Controller/FileController.php
@@ -3,11 +3,39 @@ namespace App\Controller;
 
 use App\Entity\File;
 use App\Utility\GeneralUtility;
+use App\Utility\LanguageUtility;
 
 /**
  * FileController
  */
 class FileController extends BaseController {
+    
+    /**
+     * download Action
+     * 
+     * @param \Slim\Http\Request $request
+     * @param \Slim\Http\Response $response
+     * @param array $args
+     * @return \Slim\Http\Response
+     */
+    public function download($request, $response, $args) {
+        /** @var \App\Entity\File $file **/
+        $file = $this->em->getRepository('App\Entity\File')->findOneBy(['id' => $args['uuid']]);
+        $settings = $this->container->get('settings');
+        
+        if ($file instanceof \App\Entity\File && $file->isPublic() || $file->getUser()->getId() === $this->currentUser) {
+            header("Content-Type: " . $file->getMimeType());
+            header("Content-Disposition: attachment; filename=\"" . $file->getName() . "\"");
+            flush();
+            readfile($settings['upload']['path'] . $file->getHashName() . $file->getExtension()->getName());
+            exit;
+        } else {
+            return $response->withRedirect($this->router->pathFor('file-show-' . $this->currentLocale, $args));
+        }
+        
+        // Render view
+        return $this->view->render($response, 'file/show.html.twig', array_merge($args, []));
+    }
 
     /**
      * show Action
@@ -47,7 +75,7 @@ class FileController extends BaseController {
                     $uploadFileName = $upload->getClientFilename();
                     $extension = $this->em->getRepository('App\Entity\FileExtension')->findOneBy(['name' => strtolower(substr($uploadFileName, strrpos($uploadFileName, '.'))), 'active' => 1]);
                     
-                    if ($extension instanceof \App\Entity\FileExtension) {
+                    if ($extension instanceof \App\Entity\FileExtension && $user instanceof \App\Entity\User) {
                         $uploadFileHashName = GeneralUtility::generateCode(10) . substr(md5($uploadFileName), 0, 10);
                         $upload->moveTo($settings['upload']['path'] . $uploadFileHashName . $extension->getName());
                         
@@ -60,8 +88,13 @@ class FileController extends BaseController {
                             ->setUser($user);
                         $this->em->persist($file);
                         $this->em->flush();
+                        $this->flash->addMessage('message', LanguageUtility::trans('file-upload-m1') . ';' . self::STYLE_SUCCESS);
                     }
+                } else {
+                    $this->flash->addMessage('message', LanguageUtility::trans('file-upload-m2') . ';' . self::STYLE_DANGER);
                 }
+            } else {
+                $this->flash->addMessage('message', LanguageUtility::trans('file-upload-m3') . ';' . self::STYLE_DANGER);
             }
         }
         
@@ -79,14 +112,22 @@ class FileController extends BaseController {
     public function togglePublic($request, $response, $args) {
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
         $file = $this->em->getRepository('App\Entity\File')->findOneBy(['id' => $args['uuid']]);
-        $files = $user->getFiles();
         
-        // if current user is owner of file
-        if ($files->contains($file)) {
-            $public = $file->isPublic();
-            $file->setPublic(!$public);
-            $this->em->persist($file);
-            $this->em->flush();
+        if ($user instanceof \App\Entity\User) {
+            $files = $user->getFiles();
+
+            // if current user is owner of file
+            if ($files->contains($file)) {
+                $public = $file->isPublic();
+                $file->setPublic(!$public);
+                $this->em->persist($file);
+                $this->em->flush();
+                $this->flash->addMessage('message', LanguageUtility::trans('file-public-m1') . ';' . self::STYLE_SUCCESS);
+            } else {
+                $this->flash->addMessage('message', LanguageUtility::trans('file-upload-m2') . ';' . self::STYLE_DANGER);
+            }
+        } else {
+            $this->flash->addMessage('message', LanguageUtility::trans('file-upload-m3') . ';' . self::STYLE_DANGER);
         }
         
         return $response->withRedirect($this->router->pathFor('user-show-' . $this->currentLocale));
@@ -104,13 +145,21 @@ class FileController extends BaseController {
         $settings = $this->container->get('settings');
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
         $file = $this->em->getRepository('App\Entity\File')->findOneBy(['id' => $args['uuid']]);
-        $files = $user->getFiles();
         
-        // if current user is owner of file
-        if ($files->contains($file)) {
-            unlink($settings['upload']['path'] . $file->getHashName() . $file->getExtension()->getName());
-            $this->em->remove($file);
-            $this->em->flush();
+        if ($user instanceof \App\Entity\User) {
+            $files = $user->getFiles();
+
+            // if current user is owner of file
+            if ($files->contains($file)) {
+                unlink($settings['upload']['path'] . $file->getHashName() . $file->getExtension()->getName());
+                $this->em->remove($file);
+                $this->em->flush();
+                $this->flash->addMessage('message', LanguageUtility::trans('file-remove-m1') . ';' . self::STYLE_SUCCESS);
+            } else {
+                $this->flash->addMessage('message', LanguageUtility::trans('file-remove-m2') . ';' . self::STYLE_DANGER);
+            }
+        } else {
+            $this->flash->addMessage('message', LanguageUtility::trans('file-remove-m3') . ';' . self::STYLE_DANGER);
         }
         
         return $response->withRedirect($this->router->pathFor('user-show-' . $this->currentLocale));
